Add ActiveAdmin support for exporting Excel files
This feature uses the activeadmin-axlsx gem [1]; however, the gem has
not yet be updated to support ActiveAdmin 1.0.0, so I've copied the
relevant code to the ./lib/ directory and made the necessary changes to
make it compatible with 1.0.0. This gem uses a very permissive license,
so I believe this to be fine.

When the next version of this gem comes out, the code in ./lib/ should
be removed.

[1] https://github.com/randym/activeadmin-axlsx