Bugfix - `A copy of PageCells has been removed from the module tree but is still active`
This error message would occur in development when changing a Cell
Class and reloading the associated page.

I am still a bit confused about this error and have made it my homework
to figure exactly why it was happening.

I believe this is what is happening:
 - When code is changed, the application code is reloaded on every
request (due to `config.cache_classes = false`
 - It reloads `ActiveAdmin` class which includes the `PageCells`
module. However, I believe `ActiveAdmin` is pointing to the original
`PageCells` _constant_ instead of the reloaded `PageCells` _constant_.
- Since it is pointing to a spot in memory where the the `PageCells`
module used to be, it throws up this error message.